Next.js (#26)
In order to increase the flexibility of the project, we are turning this into a `next.js` website powered by preact.

We need to add upvote features and other things in the feature. This is a step in that direction. Even though we love eleventy, it will be easier for everyone to add features to this project with React and Next.js.