Add upvote functionality (#33)
Add upvote functionality with Redis

How does the database keep track of the votes? On each Netlify build, we go to the database and **add** new projects there. Check the `bin/setup_db` script. Therefore, whenever a user requests a page, that project is already on the database. If the `setup_db` script fails, the build won't pass so users will never see projects not backed by the upvote DB.

But where is the DB?

It's on Upstash: upstash.com. It's a serverless redis platform. Much more scalable than a traditional ORM. This is also PERSISTED Redis, so our data is safe.

How do I code review this? 

Most changes are related to changing the projects json. Ignore that and the code is not that much.